+++ Parts/StartRoughDraftJob.ps1
@@ -2,20 +2,20 @@
 
 $callStackPeek = Get-PSCallStack | Select-Object -Skip 1 -First 1
 
-$InvoationInfo = $callStackPeek.InvocationInfo
+$InvocationInfo = $callStackPeek.InvocationInfo
 $myModulePath = # Determine the module path
-    if ($InvoationInfo.MyCommand.ScriptBlock.Module.Path) {
-        $psd1Path = $InvoationInfo.MyCommand.ScriptBlock.Module.Path -replace '\.psm1', '.psd1'
+    if ($InvocationInfo.MyCommand.ScriptBlock.Module.Path) {
+        $psd1Path = $InvocationInfo.MyCommand.ScriptBlock.Module.Path -replace '\.psm1', '.psd1'
         if (Test-Path $psd1Path) {
             $psd1Path
         } else {
-            $InvoationInfo.MyCommand.ScriptBlock.Module.Path
+            $InvocationInfo.MyCommand.ScriptBlock.Module.Path
         }
     } else {
         Join-Path ($psScriptRoot | Split-path) "RoughDraft.psd1"
     }
 
-$myCommandName = $InvoationInfo.MyCommand.Name # and the command name.
+$myCommandName = $InvocationInfo.MyCommand.Name # and the command name.
             $JobDefinition = [ScriptBlock]::Create({param([Hashtable]$Arguments) # Use it to create a job definition.
 }.ToString() + "
 `$myArgs = @{} + `$arguments
@@ -24,9 +24,9 @@ Set-Location `$myArgs.pwd
 `$myArgs.Remove('pwd')
 $myCommandName @myArgs
 ")
-$null= $InvoationInfo.BoundParameters.Remove("AsJob") # Remove the AsJob parameter
+$null= $InvocationInfo.BoundParameters.Remove("AsJob") # Remove the AsJob parameter
 # Pass the current directory
-$InvoationInfo.BoundParameters.PWD = "$pwd"
+$InvocationInfo.BoundParameters.PWD = "$pwd"
 # Keep track of how many jobs we've launched per command.
 if (-not $script:JobCounter) { $script:JobCounter = @{} }
 if (-not $script:JobCounter[$myCommandName]) {
@@ -37,10 +37,22 @@ if (-not $script:JobCounter[$myCommandName]) {
 
 # Launch the job and return.
 
-if ($ExecutionContext.SessionState.InvokeCommand.GetCommand('Start-ThreadJob', 'All')) {
-    Start-ThreadJob $JobDefinition -Name "$myCommandName$($script:JobCounter[$myCommandName])" -argumentlist $InvoationInfo.BoundParameters
+
+$startJobSplat = [Ordered]@{
+    ScriptBlock = $jobDefinition
+    Name = "$myCommandName$($script:JobCounter[$myCommandName])"
+    ArgumentList = $InvocationInfo.BoundParameters
+}
+
+if ($ExecutionContext.SessionState.InvokeCommand.GetCommand('Start-ThreadJob', 'All')) {    
+    if ($invocationInfo.BoundParameters.ThrottleLimit) {
+        $startJobSplat.ThrottleLimit = $invocationInfo.BoundParameters.ThrottleLimit
+    } elseif ($invocationInfo.BoundParameters.Throttle) {
+        $startJobSplat.ThrottleLimit = $invocationInfo.BoundParameters.Throttle
+    }
+    Start-ThreadJob @startJobSplat
 } else {
-    Start-Job $JobDefinition -Name "$myCommandName$($script:JobCounter[$myCommandName])" -argumentlist $InvoationInfo.BoundParameters
+    Start-Job @startJobSplat $JobDefinition -Name "$myCommandName$($script:JobCounter[$myCommandName])" -argumentlist $InvocationInfo.BoundParameters
 }
 
 return