+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-﻿## 0.3.8:
+﻿## 0.3.9:
+
+* New Extensions:
+  * ColorExpression ( Fixes #172 )
+  * MixVideo ( Fixes #171 )
+* Extension Updates:
+  * BlendFrame now supports by Join-Media (Fixes #173)
+* All Valid -Media Commands now support -FFMpegArgument (Fixes #75)
+* Better Jobs Support
+  * Join-Media and New-Media now support -AsJob (Fixes #174)
+  * Job commands now support -ThrottleLimit (Fixes #175)
+
+---
+
+## 0.3.8:
 
 * New Extensions:
   * AudioFrequencyExpression (Fixes #158)